requirements.txt
vk_api
pyTelegramBotAPI
requests
pyyaml
